Move txn db function to new crate
